requirements.txt
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.3
mysql-connector-python>=8.0.28
pyodbc>=4.0.34
pytest>=7.0.0
python-dotenv>=1.0.0
fastmcp>=2.0.0
